requirements.txt
asgiref==3.9.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2.5
djangorestframework==3.16.1
idna==3.10
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.4.0